+++ Modules/booking/Controller/BookingauthorisationsController.php
@@ -30,7 +30,7 @@ class BookingauthorisationsController extends CoresecureController {
 
     public function indexAction($id_space, $id) {
 
-        $this->checkAuthorizationMenuSpace("ecusers", $id_space, $_SESSION["id_user"]);
+        $this->checkSpaceAdmin($id_space, $_SESSION["id_user"]);
         $lang = $this->getLanguage();
 
         // get all the resources
@@ -95,7 +95,7 @@ class BookingauthorisationsController extends CoresecureController {
 
     public function historyAction($id_space, $id) {
 
-        $this->checkAuthorizationMenuSpace("ecusers", $id_space, $_SESSION["id_user"]);
+        $this->checkSpaceAdmin($id_space, $_SESSION["id_user"]);
         $lang = $this->getLanguage();
 
         $idArray = explode("_", $id);
@@ -156,7 +156,7 @@ class BookingauthorisationsController extends CoresecureController {
     }
 
     public function addAction($id_space, $id) {
-        $this->checkAuthorizationMenuSpace("ecusers", $id_space, $_SESSION["id_user"]);
+        $this->checkSpaceAdmin($id_space, $_SESSION["id_user"]);
         $lang = $this->getLanguage();
 
         $idArray = explode("_", $id);
@@ -214,7 +214,7 @@ class BookingauthorisationsController extends CoresecureController {
     }
 
     public function editAction($id_space, $id) {
-        $this->checkAuthorizationMenuSpace("ecusers", $id_space, $_SESSION["id_user"]);
+        $this->checkSpaceAdmin($id_space, $_SESSION["id_user"]);
         $lang = $this->getLanguage();
 
         $modelAuth = new BkAuthorization();
@@ -226,6 +226,7 @@ class BookingauthorisationsController extends CoresecureController {
 
         $modelResourcesCategories = new ReCategory();
         // $categoryName = $modelResourcesCategories->getName($data["resource_id"]);
+        $id_resource_category = $data["resource_id"];
         $recat = $modelResourcesCategories->get($id_space, $id_resource_category);
         $categoryName = $recat['name'];