+++ Modules/catalog/Controller/CatalogadminController.php
@@ -29,6 +29,7 @@ class CatalogadminController extends CoresecureController {
      * @see Controller::categoriesAction()
      */
     public function categoriesAction($id_space) {
+        $this->checkAuthorizationMenuSpace("catalogsettings", $id_space, $_SESSION["id_user"]);
         $lang = $this->getLanguage();
         // get the user list
         $modelCategory = new CaCategory();
@@ -44,6 +45,7 @@ class CatalogadminController extends CoresecureController {
         ));
     }
     public function categoryeditAction($id_space, $id) {
+        $this->checkAuthorizationMenuSpace("catalogsettings", $id_space, $_SESSION["id_user"]);
         $lang = $this->getLanguage();
         // get name
         $name = "";
@@ -82,12 +84,14 @@ class CatalogadminController extends CoresecureController {
      * Remove a category from the database
      */
     public function categorydeleteAction($id_space, $id) {
+        $this->checkAuthorizationMenuSpace("catalogsettings", $id_space, $_SESSION["id_user"]);
         $modelCategory = new CaCategory();
         $modelCategory->delete($id_space, $id);
         // generate view
         $this->redirect("catalogcategories/".$id_space);
     }
     public function prestationsAction($id_space) {
+        $this->checkAuthorizationMenuSpace("catalogsettings", $id_space, $_SESSION["id_user"]);
         $lang = $this->getLanguage();
         // get the user list
         $modelEntry = new CaEntry();
@@ -110,6 +114,7 @@ class CatalogadminController extends CoresecureController {
         ));
     }
     public function prestationeditAction($id_space, $id) {
+        $this->checkAuthorizationMenuSpace("catalogsettings", $id_space, $_SESSION["id_user"]);
         $lang = $this->getLanguage();
         // get info
         $modelEntry = new CaEntry();
@@ -146,10 +151,9 @@ class CatalogadminController extends CoresecureController {
             } else {
                 $id = $modelEntry->add($id_space, $id_category, $title, $short_desc, $full_desc);
             }
-            //echo "file = " . $_FILES["illustration"]["name"] . "<br/>";
             if ($_FILES["illustration"]["name"] != "") {
-                // download file
-                $this->downloadIllustration();
+                // upload file
+                $this->uploadIllustration();
                 // set filename to database
                 $modelEntry->setImageUrl($id_space, $id, $_FILES["illustration"]["name"]);
             }
@@ -164,30 +168,29 @@ class CatalogadminController extends CoresecureController {
             ));
         }
     }
-    protected function downloadIllustration() {
+    protected function uploadIllustration() {
         $target_dir = "data/catalog/";
         $target_file = $target_dir . $_FILES["illustration"]["name"];
-        //echo "target file = " . $target_file . "<br/>";
-        $uploadOk = 1;
-        //$imageFileType = pathinfo($_FILES["illustration"]["name"],PATHINFO_EXTENSION);
+
+        $fileNameOK = preg_match("/^[0-9a-zA-Z\-_\.]+$/", $_FILES["illustration"]["name"], $matches);
+        if(! $fileNameOK) {
+            throw new PfmFileException("invalid file name, must be alphanumeric:  [0-9a-zA-Z\-_\.]+", 403);
+        }
+
         // Check file size
         if ($_FILES["illustration"]["size"] > 500000000) {
             return "Error: your file is too large.";
-            //$uploadOk = 0;
         }
-        // Check if $uploadOk is set to 0 by an error
-        if ($uploadOk == 0) {
-            return "Error: your file was not uploaded.";
-            // if everything is ok, try to upload file
+        
+        if (move_uploaded_file($_FILES["illustration"]["tmp_name"], $target_file)) {
+            return "The file logo file" . basename($_FILES["illustration"]["name"]) . " has been uploaded.";
         } else {
-            if (move_uploaded_file($_FILES["illustration"]["tmp_name"], $target_file)) {
-                return "The file logo file" . basename($_FILES["illustration"]["name"]) . " has been uploaded.";
-            } else {
-                return "Error, there was an error uploading your file.";
-            }
+            return "Error, there was an error uploading your file.";
         }
     }
+
     public function prestationdeleteAction($id_space, $id) {
+        $this->checkAuthorizationMenuSpace("catalogsettings", $id_space, $_SESSION["id_user"]);
         $modelCategory = new CaEntry();
         $modelCategory->delete($id_space ,$id);
         // generate view