+++ Modules/core/Controller/CorespaceadminController.php
@@ -10,7 +10,6 @@ require_once 'Framework/Form.php';
 require_once 'Framework/FileUpload.php';
 require_once 'Modules/core/Controller/CoresecureController.php';
 require_once 'Modules/core/Model/CoreInstall.php';
-require_once 'Modules/core/Model/CoreBackupDatabase.php';
 require_once 'Modules/core/Model/CoreSpace.php';
 
 require_once 'Modules/core/Model/CoreUser.php';