+++ Modules/services/Controller/ServiceslistingController.php
@@ -80,7 +80,7 @@ class ServiceslistingController extends CoresecureController {
 
         $form->addSelect("type_id", CoreTranslator::type($lang), $types["names"], $types["ids"], $value["type_id"]);
         $form->setValidationButton(CoreTranslator::Save($lang), "servicesedit/" . $id_space . "/" . $id);
-        $form->setCancelButton(CoreTranslator::Cancel($lang), "services/" . $id_space);
+        $form->setCancelButton(CoreTranslator::Cancel($lang), "serviceslisting/" . $id_space);
 
         if ($form->check()) {
             $this->serviceModel->setService(
@@ -102,7 +102,7 @@ class ServiceslistingController extends CoresecureController {
         $this->checkAuthorizationMenuSpace("services", $id_space, $_SESSION["id_user"]);
 
         $this->serviceModel->delete($id_space, $id);
-        $this->redirect("services/" . $id_space);
+        $this->redirect("serviceslisting/" . $id_space);
     }
 
     public function stockAction($id_space) {