+++ CHANGELOG.md
@@ -2,13 +2,33 @@
 
 ## 2.2
 
+<<<<<<< HEAD
 * [module:quote] expose client infos to invoice templates
+=======
+* [Users] Remove login from edit function
+* [self_registration] fix selfregistration email sending order
+* [Security] Add missing access authorization controls
+>>>>>>> pfm2
 * [Users] Improve users creation forms controls
 * [resources] set default booking authorizations at resource creation
 * [module:booking_settings] fix display edition
 * [stats] count number of tickets per status
 * [stats] add stat calentry_cancel on booking cancel
 * [mail] allow users to unsubscribe to notifications #382
+* [core][ldap] rename base config parameters for ldap auth
+* [helpdesk] ignore delivery status notifications (do not reply)
+* [helpdesk] let user select multiple tickets to spam them #393
+* [core] limit file uploads name to alphanumeric # 402
+* [core] add welcome page and use it as default entry url
+
+## 2.1.9
+
+* [users] fix users_info sql
+* [helpdesk] check if message is an auto-reply and log
+* [booking][invoice] fix sql
+* [booking][colorcode] fix sql
+* [sql] fix wrong space_id error => id_space (param and sql)
+* [booking][calsup] fix remove unlisted supinfo
 
 ## 2.1.8